Don't compile SDK if SDK_SRC_PATH is not defined or not a valid path
git-svn-id: file:///afs/slac/g/reseng/svn/repos@12369 0848cbd7-7569-0410-ab05-8cc0a0c67708

Former-commit-id: b176ff68205993282896565771007935cc4ae040
Former-commit-id: 9ad743218aa994f2fb50594d1952fb3ea47c859d
Former-commit-id: 0411cd5bb398600d516c48ef7e385617c71afc6c

+++ StdLib/trunk/build/vivado_post_route_v1.tcl
@@ -40,27 +40,31 @@ if { [CheckTiming false] == true } {
    #########################################################
    ## Check if SDK's .sysdef file exists
    #########################################################
-   if { [file exists ${OUT_DIR}/${VIVADO_PROJECT}.runs/impl_1/${PROJECT}.sysdef] == 1 } {
-      # Check if custom SDK exist
-      if { [file exists ${VIVADO_DIR}/sdk.tcl] == 1 } {   
-         source ${VIVADO_DIR}/sdk.tcl
-      } else {
-         set SDK_PRJ_RDY false
-         while { ${SDK_PRJ_RDY} != true } {
-            set src_rc [catch {exec xsdk -batch -source ${VIVADO_BUILD_DIR}/vivado_sdk_prj_v1.tcl >@stdout}]       
-            if {$src_rc} {
-               puts "Retrying to build SDK project"
-               exec rm -rf ${SDK_PRJ}
-            } else {
-               set SDK_PRJ_RDY true
-            }         
-         }
-         # Try to build the .ELF file
-         catch { 
-            # Generate .ELF
-            set src_rc [catch {exec xsdk -batch -source ${VIVADO_BUILD_DIR}/vivado_sdk_elf_v1.tcl >@stdout}]    
-            # Add .ELF to the .bit file
-            source ${VIVADO_BUILD_DIR}/vivado_sdk_bit_v1.tcl       
+   # Check if SDK_SRC_PATH is a valid path
+   if { [CheckSdkSrcPath] != false } {
+      # Check for .sysdef file (generated when using Microblaze)
+      if { [file exists ${OUT_DIR}/${VIVADO_PROJECT}.runs/impl_1/${PROJECT}.sysdef] == 1 } {
+         # Check if custom SDK exist
+         if { [file exists ${VIVADO_DIR}/sdk.tcl] == 1 } {   
+            source ${VIVADO_DIR}/sdk.tcl
+         } else {
+            set SDK_PRJ_RDY false
+            while { ${SDK_PRJ_RDY} != true } {
+               set src_rc [catch {exec xsdk -batch -source ${VIVADO_BUILD_DIR}/vivado_sdk_prj_v1.tcl >@stdout}]       
+               if {$src_rc} {
+                  puts "Retrying to build SDK project"
+                  exec rm -rf ${SDK_PRJ}
+               } else {
+                  set SDK_PRJ_RDY true
+               }         
+            }
+            # Try to build the .ELF file
+            catch { 
+               # Generate .ELF
+               set src_rc [catch {exec xsdk -batch -source ${VIVADO_BUILD_DIR}/vivado_sdk_elf_v1.tcl >@stdout}]    
+               # Add .ELF to the .bit file
+               source ${VIVADO_BUILD_DIR}/vivado_sdk_bit_v1.tcl       
+            }
          }
       }
    }
